refactor: improvement setup e2e test
